Files Added: - BodyContent.jsx <- houses the game cards - Footer.jsx <- footer, as the name states - GameCard.jsx <- gets the data for the games, then makes them into renderable cards - Hub.jsx <- houses all of the program stuff, aka the main react module - NavBar.jsx <- navigation bar
Added the basic react layout, has:
- header
- body content
- footer

Need to fix data I think, since I just added sample data into the game cards so it doesn't create a new one inside of the container each time. Meaning the render out into a single vertical line of games. But I tested with 2 games and it seems fine, the css also sorta scales with the screensize but I have to test this more